refactor: as much as possible, move Kani lib exe trace code to separate module
