+++ library/kani/src/lib.rs
@@ -14,25 +14,19 @@
 #![feature(generic_const_exprs)]
 
 pub mod arbitrary;
+#[cfg(feature = "exe_trace")]
+pub mod exe_trace;
 pub mod futures;
 pub mod invariant;
 pub mod slice;
 pub mod vec;
 
 pub use arbitrary::Arbitrary;
+#[cfg(feature = "exe_trace")]
+pub use exe_trace::exe_trace_run;
 pub use futures::block_on;
 #[allow(deprecated)]
 pub use invariant::Invariant;
-
-#[cfg(feature = "exe_trace")]
-use std::sync::Mutex;
-/// DET_VALS_LOCK is used by each playback test case to ensure that only a single thread is modifying DET_VALS at once.
-/// We need to separate the lock from the data because there's no other way to pass the data from
-/// kani::exe_trace_run() to kani::any_raw_internal() while still holding the lock.
-#[cfg(feature = "exe_trace")]
-static DET_VALS_LOCK: Mutex<()> = Mutex::new(());
-#[cfg(feature = "exe_trace")]
-static mut DET_VALS: Vec<Vec<u8>> = Vec::new();
 
 /// Creates an assumption that will be valid after this statement run. Note that the assumption
 /// will only be applied for paths that follow the assumption. If the assumption doesn't hold, the
@@ -136,7 +130,7 @@ pub(crate) unsafe fn any_raw_internal<T, const SIZE_T: usize>() -> T {
     #[cfg(feature = "exe_trace")]
     {
         // This code will only run when our thread's exe_trace_run() fn holds the lock.
-        let next_det_val = DET_VALS.pop().expect("Not enough det vals found");
+        let next_det_val = exe_trace::DET_VALS.pop().expect("Not enough det vals found");
         let next_det_val_len = next_det_val.len();
         let bytes_t: [u8; SIZE_T] = next_det_val.try_into().expect(&format!(
             "Expected {SIZE_T} bytes instead of {next_det_val_len} bytes in the following det vals vec"
@@ -155,25 +149,6 @@ pub(crate) unsafe fn any_raw_internal<T, const SIZE_T: usize>() -> T {
 #[allow(dead_code)]
 fn any_raw_inner<T>() -> T {
     unimplemented!("Kani any_raw_inner");
-}
-
-/// This function sets deterministic values and plays back the user's proof harness.
-#[cfg(feature = "exe_trace")]
-pub fn exe_trace_run<F: Fn()>(mut det_vals: Vec<Vec<u8>>, proof_harness: F) {
-    // Det vals in the user test case should be in the same order as the order of kani::any() calls.
-    // Here, we need to reverse this order because det vals are popped off of the outer Vec,
-    // so the chronological first det val should come last.
-    det_vals.reverse();
-    // If another thread panicked while holding the lock (e.g., because they hit an expected assertion failure), we still want to continue.
-    let _guard = match DET_VALS_LOCK.lock() {
-        Ok(guard) => guard,
-        Err(poisoned) => poisoned.into_inner(),
-    };
-    unsafe {
-        DET_VALS = det_vals;
-    }
-    // Since F is a type argument, there should be a direct, static call to proof_harness().
-    proof_harness();
 }
 
 /// Function used in tests for cases where the condition is not always true.
